refactor: delete layer_linking, move to resource_linking
