--- tests/unit/hook_packages/terraform/test_layer_linking.py
@@ -1,67 +1,0 @@
-from copy import deepcopy
-from unittest import TestCase
-
-from parameterized import parameterized
-
-from samcli.hook_packages.terraform.hooks.prepare.layer_linking import _clean_references_list
-
-
-class TestLayerLinking(TestCase):
-    @parameterized.expand(
-        [
-            ([], []),
-            (["aws_lambda_layer_version.layer1[0].arn"], ["aws_lambda_layer_version.layer1[0].arn"]),
-            (["aws_lambda_layer_version.layer1[0]"], ["aws_lambda_layer_version.layer1[0]"]),
-            (["aws_lambda_layer_version.layer1"], ["aws_lambda_layer_version.layer1"]),
-            (
-                [
-                    "aws_lambda_layer_version.layer1[0].arn",
-                    "aws_lambda_layer_version.layer1[0]",
-                    "aws_lambda_layer_version.layer1",
-                ],
-                ["aws_lambda_layer_version.layer1[0].arn"],
-            ),
-            (
-                [
-                    "aws_lambda_layer_version.layer1[0].arn",
-                    "aws_lambda_layer_version.layer1[0]",
-                    "aws_lambda_layer_version.layer1",
-                    "module.const_layer1.layer_arn",
-                    "module.const_layer1",
-                    "module.const_layer2.layer_arn",
-                    "module.const_layer2",
-                ],
-                [
-                    "module.const_layer2.layer_arn",
-                    "module.const_layer1.layer_arn",
-                    "aws_lambda_layer_version.layer1[0].arn",
-                ],
-            ),
-            (
-                [
-                    'aws_lambda_layer_version.layer1["key1"].arn',
-                    'aws_lambda_layer_version.layer1["key1"]',
-                    "aws_lambda_layer_version.layer1",
-                ],
-                ['aws_lambda_layer_version.layer1["key1"].arn'],
-            ),
-        ]
-    )
-    def test_clean_references_list(self, references, expected):
-        cleaned_references = _clean_references_list(references)
-        self.assertEqual(cleaned_references, expected)
-
-    def test_ensure_original_references_not_mutated(self):
-        references = [
-            "aws_lambda_layer_version.layer1[0].arn",
-            "aws_lambda_layer_version.layer1[0]",
-            "aws_lambda_layer_version.layer1",
-            "module.const_layer1.layer_arn",
-            "module.const_layer1",
-            "module.const_layer2.layer_arn",
-            "module.const_layer2",
-        ]
-        original_references = deepcopy(references)
-        cleaned_references_list = _clean_references_list(references)
-        self.assertEqual(references, original_references)
-        self.assertNotEqual(references, cleaned_references_list)
